chore: authStore.ts 삭제
